Rename project.
