+++ wasm_interface/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "jni_interface"
+name = "wasm_interface"
 version = "0.1.0"
 edition = "2018"
