+++ wasm_interface/src/allocator.rs
